added support for custom firefox args and preferences
