+++ carina-core/src/main/resources/config.properties
@@ -32,7 +32,8 @@ browsermob_ports_range=NULL
 # browser options and arguments
 chrome_args=NULL
 chrome_experimental_opts=NULL
-
+firefox_args=NULL
+firefox_preferences=NULL
 
 #L10N - ENV_LOCALE
 enable_l10n=true
